Time: 21 ms (72.27%) | Memory: 63.8 MB (76.30%) - LeetSync

+++ 1306-minimum-absolute-difference/minimum-absolute-difference.java
@@ -2,7 +2,7 @@ class Solution {
     public List<List<Integer>> minimumAbsDifference(int[] arr) {
     
         int n = arr.length;
-        long minDiff = Long.MAX_VALUE;
+        int minDiff = Integer.MAX_VALUE;
         Arrays.sort(arr);
         List<List<Integer>>  ans  = new ArrayList<>();
         for(int i=0 ;i<n-1 ;i++){
